+++ src/check/Applications.ps1
@@ -367,10 +367,7 @@ function Invoke-RootFolderPermissionCheck {
     )
 
     begin {
-        # $IgnoredRootFolders = @( "Windows", "Users", "Program Files", "Program Files (x86)", "PerfLogs")
         $IgnoredRootFolders = @( "`$Recycle.Bin", "`$WinREAgent", "Documents and Settings", "PerfLogs", "Program Files", "Program Files (x86)", "ProgramData", "Recovery", "System Volume Information", "Users", "Windows", "Windows.old" )
-        # $MaxFileCount = 8
-        # $AllResults = @()
     }
 
     process {
@@ -401,7 +398,7 @@ function Invoke-RootFolderPermissionCheck {
             Write-Progress -Activity "Checking root folder permissions ($($Candidates.Count)/$($Candidates.Count))..." -Status "100% Complete:" -Completed
         }
 
-        $Vulnerable = ($Results | Where-Object { $_.Vulnerable }).Count -gt 0
+        $Vulnerable = ([object[]]($Results | Where-Object { $_.Vulnerable })).Count -gt 0
 
         $CheckResult = New-Object -TypeName PSObject
         $CheckResult | Add-Member -MemberType "NoteProperty" -Name "Result" -Value $Results